refactor header into a component, sort by latest post on home screen, export components from folder
