+++ src/components/Hero.svelte
@@ -1,0 +1,30 @@
+<h1>Recent articles</h1>
+
+<figure>
+  <figcaption>The most recent trends in Web Devops practices!</figcaption>
+</figure>
+<style>
+    
+  h1,
+  figure {
+    text-align: center;
+    margin: 0 auto;
+  }
+
+  h1 {
+    font-size: 2.8em;
+    text-transform: uppercase;
+    font-weight: 700;
+    margin: 0 0 0.5em 0;
+  }
+
+  figure {
+    margin: 0 0 2em 0;
+  }
+
+  @media (min-width: 480px) {
+    h1 {
+      font-size: 4em;
+    }
+  }
+</style>
