+++ src/components/IntersectionObserver.svelte
