+++ src/components/index.js
@@ -1,0 +1,5 @@
+export { default as BlogCard } from './BlogCard.svelte'
+export { default as Hero } from './Hero.svelte'
+export { default as IntersectionObserver } from './IntersectionObserver.svelte'
+export { default as Nav } from './Nav.svelte'
+export { default as SocialShare } from './SocialShare.svelte'
